requirements.txt
aiohttp==3.8.5
asyncio==3.4.3
fastapi==0.104.1
pydantic==2.4.2
pydantic[email]==2.4.2
pytz==2023.3
python-dateutil==2.8.2
python-dotenv==1.0.0
supabase==1.0.4
gunicorn==21.2.0
agents==0.3.0
calendar==0.5.0
uuid==1.30 